requirements.txt
pelican[markdown]~=4.11
pelican-seo~=1.4
pelican-sitemap~=1.2
pelican-htmlmin~=2.0